==> 3. Train & validation data split.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load in 

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the "../input/" directory.
# For example, running this (by clicking run or pressing Shift+Enter) will list the files in the input directory

import os
print(os.listdir("../input"))

# Any results you write to the current directory are saved as output.

import pickle
import random
# 5000 train - 3G
def train_val_split(file_dir):
    # read data
    f = open(file_dir + 'article_embedding_pad.pickle','rb')
    article_embedding = pickle.load(f)
    f.close()
    f = open(file_dir + 'title_embedding_pad.pickle','rb')
    title_embedding = pickle.load(f)
    f.close()
    f = open(file_dir + 'train_label_list.pickle','rb')
    label_list = pickle.load(f)
    f.close()
    # split validation 5000 samples
    article_id = [i for i in range(len(article_embedding))]
    random.seed(0)
    random.shuffle(article_id)
    val_id = article_id[:5000]
    train_id = article_id[5000:]
    train_article_embedding_data = [article_embedding[i] for i in train_id]
    train_title_embedding_data = [title_embedding[i] for i in train_id]
    train_label_list_data = [label_list[i] for i in train_id]
    val_article_embedding_data = [article_embedding[i] for i in val_id]
    val_title_embedding_data = [title_embedding[i] for i in val_id]
    val_label_list_data = [label_list[i] for i in val_id]
    # create vocal label 
    tmp_label1 = train_label_list_data.copy()
    tmp_label2 = val_label_list_data.copy()
    tmp_label1.extend(tmp_label2)
    tmp = {}
    for word in tmp_label1:
        for i in word:
            tmp[i] = tmp.get(i,0) + 1
    label_vocab = {}
    index = 0
    for key in tmp:
        if tmp[key] > 10:
            label_vocab[key] = index
            index += 1
    return train_article_embedding_data, train_title_embedding_data, train_label_list_data,\
           val_article_embedding_data, val_title_embedding_data, val_label_list_data, label_vocab

def one_hot_label_embedding(label_list, label_vocab, start_id, end_id):
    tmp = label_list[start_id:end_id+1]
    tmp_embed = []
    for i in tmp:
        l = [0] * len(label_vocab)
        for j in i:
            l[label_vocab[j]] = 1
        tmp_embed.append(l)
    return tmp_embed

train_article_embedding_data, train_title_embedding_data, train_label_list_data,\
val_article_embedding_data, val_title_embedding_data, val_label_list_data, label_vocab = train_val_split('../input/')

f = open('train_article_embedding_data.pickle','wb')
pickle.dump(train_article_embedding_data, f)
f.close()
f = open('train_title_embedding_data.pickle','wb')
pickle.dump(train_title_embedding_data, f)
f.close()
f = open('train_label_list_data.pickle','wb')
pickle.dump(train_label_list_data, f)
f.close()
f = open('val_article_embedding_data.pickle','wb')
pickle.dump(val_article_embedding_data, f)
f.close()
f = open('val_title_embedding_data.pickle','wb')
pickle.dump(val_title_embedding_data, f)
f.close()
f = open('val_label_list_data.pickle','wb')
pickle.dump(val_label_list_data, f)
f.close()
f = open('label_vocab.pickle','wb')
pickle.dump(label_vocab, f)
f.close()
