==> 6. Create Prediction File.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load in 

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the "../input/" directory.
# For example, running this (by clicking run or pressing Shift+Enter) will list the files in the input directory

import os
print(os.listdir("../input"))

# Any results you write to the current directory are saved as output.

import numpy as np # linear algebra
import os
print(os.listdir("../input"))
import pickle
import logging
logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s',level=logging.INFO)
args = {}
def test_read_input(fold_dir):
    raw_data = pd.read_csv(fold_dir+'test.csv')
    logging.info("reading raw data...this may take a while")
    l = len(raw_data)
    id_list = []
    for i in range(l):
        id = raw_data['id'].iloc[i]
        if (i % 10000 == 0):
            logging.info("read {0} reviews".format(i))
        id_list.append(id)
    return id_list

test_id_list = test_read_input('../input/hackerearth-deep-learning-challenge-4/')

def read_vocab(fold_dir,args):        
    f = open(fold_dir + 'label_vocab.pickle','rb')
    args['label_vocab'] = pickle.load(f)
    f.close()
    args['label_dim'] = len(args['label_vocab'])
    f.close()
    return args

args = read_vocab('../input/tag-recommendation-system-list/', args)

#f = open('../input/fork-of-prediction-for-test-data-1/pred_y1.pickle','rb')
#pred_y = pickle.load(f)
#f.close()
result = []
args['id_to_label'] = {args['label_vocab'][i] : i for i in args['label_vocab']}
output = open('predict.csv','w')
output.write('id,tags\n')
id_i = 0
for i in range(1,10):
    f = open('../input/fork-of-prediction-for-test-data-'+ str(i) +'/pred_y' + str(i) + '_all.pickle','rb')
    pred_y = pickle.load(f)
    print(len(pred_y))
    f.close()
    for line_id in range(len(pred_y)):
        if line_id % 5000 == 0:
            print(line_id,'finish\n')
        y = pred_y[line_id]
        label_y = []
        for i in range(len(y)):
            if y[i] == 1:
                label_y.append(args['id_to_label'][i])
        label_str='|'.join(label_y)
        output.write('%s,%s\n' % (test_id_list[id_i], label_str))
        id_i += 1
    del pred_y
output.close()

'''
def create_predict_file(args,test_id_list,result):
    # args['id_to_label'] = {args['label_vocab'][i] : i for i in args['label_vocab']}
    f = open('predict.csv','w')
    f.write('id,tags\n')
    for line_id in range(len(result)):
        f.write('%s,%s\n' % (test_id_list[line_id], result[line_id]))
    f.close()
    return

create_predict_file(args,test_id_list,result)
'''
